deleted old App.jsx and renamed AppCopy.jsx to App.jsx/
